source_code/requirements.txt
google-cloud-bigquery==2.34.0